main/infra/github/pyproject.toml
[tool.poetry]
name = "runner"
version = "0.1.0"
description = ""
authors = ["James Winegar <jameswine@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
PyGithub = "^1.59"
psycopg = {extras = ["binary"], version = "^3.1.10"}
python-dotenv = "^1.0.0"
pandas = "^2.0.3"
sqlalchemy = "^2.0.20"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.286"
pytest = "^7.4"
black = "^23.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = "runner.main:runner"
